Fix an error from undefined behavior sanitizer in xz_config.h
A cast to uint32_t was missing, thus the left shift was done on
a signed int (promoted from uint8_t). The shift could touch the
sign bit which is undefined behavior.

xz_config.h is for userspace builds; it's not in the Linux kernel.

+++ userspace/xz_config.h
@@ -96,7 +96,7 @@ static inline uint32_t get_unaligned_le32(const uint8_t *buf)
 #ifndef get_unaligned_be32
 static inline uint32_t get_unaligned_be32(const uint8_t *buf)
 {
-	return (uint32_t)(buf[0] << 24)
+	return (uint32_t)((uint32_t)buf[0] << 24)
 			| ((uint32_t)buf[1] << 16)
 			| ((uint32_t)buf[2] << 8)
 			| (uint32_t)buf[3];
